Move laws classes out of entity package
