+++ src/main/java/com/example/demo/models/businesslogic/Law.java
@@ -1,5 +1,8 @@
-package com.example.demo.models.entities;
+package com.example.demo.models.businesslogic;
 
+import com.example.demo.models.entities.Concept;
+import com.example.demo.models.entities.LawFormulation;
+import com.example.demo.models.entities.Tag;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
