+++ src/main/java/com/example/demo/models/businesslogic/QuestionRequest.java
@@ -1,10 +1,8 @@
 package com.example.demo.models.businesslogic;
 
 import com.example.demo.models.entities.Concept;
-import com.example.demo.models.entities.Law;
 import lombok.Data;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Data
