+++ src/main/java/com/example/demo/models/businesslogic/backend/Backend.java
@@ -1,6 +1,6 @@
 package com.example.demo.models.businesslogic.backend;
 
-import com.example.demo.models.entities.Law;
+import com.example.demo.models.businesslogic.Law;
 import com.example.demo.models.entities.Mistake;
 
 import java.util.List;
